added file upload and pdf generation

+++ controllers/admin.js
@@ -1,3 +1,6 @@
+const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -15,12 +18,30 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
+  // console.log(image);
+
+  if (!image) {
+    res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMessage: 'Please select an image file of jpg, jpeg or png format.',
+      validationErrors: []
+    });
+  }
 
   if (!errors.isEmpty()) {
+    // console.log(errors);
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
@@ -28,7 +49,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description
       },
@@ -36,6 +56,8 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
@@ -52,11 +74,11 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      // console.log(err);
-      
+      console.log(err);
+
       // res.redirect('/500');
 
-      const error =  new Error(err);
+      const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
     });
@@ -88,7 +110,7 @@ exports.getEditProduct = (req, res, next) => {
 
       // res.redirect('/500');
 
-      const error =  new Error(err);
+      const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
     });
@@ -98,7 +120,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -129,7 +151,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save()
         .then(result => {
           // console.log('UPDATED PRODUCT!');
@@ -139,7 +164,7 @@ exports.postEditProduct = (req, res, next) => {
     .catch(err => {
       // console.log(err);
       // res.redirect('/500');
-      const error =  new Error(err);
+      const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
     });
@@ -160,7 +185,7 @@ exports.getProducts = (req, res, next) => {
     })
     .catch(err => {
       // console.log(err);
-      const error =  new Error(err);
+      const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
     });
@@ -168,15 +193,22 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       // console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
     .catch(err => {
       // console.log(err);
-      const error =  new Error(err);
+      const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
-    });
+    });    
 };
